fix: check whether items are already archived in bulk actions dropdown
