--- apps/hash-frontend/src/shared/entity-types-context/util.ts
@@ -1,14 +1,0 @@
-import {
-  DataTypeWithMetadata,
-  EntityTypeWithMetadata,
-  PropertyTypeWithMetadata,
-} from "@local/hash-subgraph";
-
-export const isTypeArchived = (
-  type:
-    | EntityTypeWithMetadata
-    | PropertyTypeWithMetadata
-    | DataTypeWithMetadata,
-) =>
-  type.metadata.custom.temporalVersioning.transactionTime.end.kind ===
-  "exclusive";
