+++ apps/hash-frontend/src/shared/util.ts
@@ -5,6 +5,7 @@ import {
   EntityTypeWithMetadata,
   PropertyTypeWithMetadata,
 } from "@local/hash-subgraph";
+import { extractBaseUrl } from "@local/hash-subgraph/type-system-patch";
 
 export const isType = (
   item:
@@ -33,3 +34,29 @@ export const isTypePropertyType = (
 
 export const isEntityPageEntity = (item: Entity) =>
   item.metadata.entityTypeId === types.entityType.page.entityTypeId;
+
+export const isTypeArchived = (
+  type:
+    | EntityTypeWithMetadata
+    | PropertyTypeWithMetadata
+    | DataTypeWithMetadata,
+) =>
+  type.metadata.custom.temporalVersioning.transactionTime.end.kind ===
+  "exclusive";
+
+export const isItemArchived = (
+  item:
+    | Entity
+    | EntityTypeWithMetadata
+    | PropertyTypeWithMetadata
+    | DataTypeWithMetadata,
+) => {
+  if (isType(item)) {
+    return isTypeArchived(item);
+  } else if (isEntityPageEntity(item)) {
+    return item.properties[
+      extractBaseUrl(types.propertyType.archived.propertyTypeId)
+    ] as boolean;
+  }
+  return false;
+};
